requirements.txt


streamlit
openai
python-dotenv
google-generativeai